+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 <!-- markdownlint-disable MD041 -->
 
+## Unreleased
+
+- added
+  - taxel-py: Generate xml from xsd schema
+  - Add CI
+- changed
+- removed
+
 ## v0.1.0 (2022-06-11)
 
 Initial release